+++ elements/drawtools/stories/multi-feature-select.js
@@ -11,7 +11,8 @@ export const MuliFeatureSelect = {
     type: "Box",
     layerId: "regions",
     showList: true,
-    featureName: "Selection",
+    featureName: "Region",
+    featureNameKey: "ECO_NAME",
     featureStyles: {
       layer: {
         "fill-color": "#16A105A0",
@@ -23,27 +24,22 @@ export const MuliFeatureSelect = {
         "stroke-color": "#19B806",
         "stroke-width": 2.5,
       },
-      click: {
-        "fill-color": "#1FD609A0",
-        "stroke-color": "#1FD609",
-        "stroke-width": 2.5,
-      },
     },
     drawUpdate: (e) => {
       console.log("drawUpdate:", e.detail);
     },
   },
   render: (args) => html`
-    <!-- Render eox-map component with ID "primary" -->
+    <!-- Render eox-map component with ID "multi-select" -->
     <eox-map
-      id="primary"
+      id="multi-select"
       style=${STORIES_MAP_STYLE}
       .layers=${STORIES_VECTOR_LAYERS}
     ></eox-map>
 
-    <!-- Initialize eox-drawtools for the eox-map with ID "primary" -->
+    <!-- Initialize eox-drawtools for the eox-map with ID "multi-select" -->
     <eox-drawtools
-      for="eox-map#primary"
+      for="eox-map#multi-select"
       .allowModify=${args.allowModify}
       .multipleFeatures=${args.multipleFeatures}
       .type=${args.type}
@@ -51,6 +47,7 @@ export const MuliFeatureSelect = {
       ?show-list=${args.showList}
       .featureStyles=${args.featureStyles}
       .featureName=${args.featureName}
+      .featureNameKey=${args.featureNameKey}
       @drawupdate=${args.drawUpdate}
     />
   `,