+++ elements/tour/stories/index.js
@@ -2,3 +2,4 @@
 // Export different stories
 
 export { default as PrimaryStory } from "./primary";
+export { default as TourLoadOnceStory } from "./tour-load-once";