+++ elements/tour/stories/tour.stories.js
@@ -2,7 +2,7 @@
  * Stories for eox-tour component showcasing various configurations.
  * These stories provide visual representations and usage examples for different scenarios.
  */
-import { PrimaryStory } from "./index";
+import { PrimaryStory, TourLoadOnceStory } from "./index";
 
 export default {
   title: "Elements/eox-tour",
@@ -16,3 +16,8 @@ export default {
  * Primary story showcasing basic usage.
  */
 export const Primary = PrimaryStory;
+
+/**
+ * Story in which the tour is loaded once and if the user delete the tour state the tour is loaded again.
+ */
+export const TourLoadOnce = TourLoadOnceStory;