+++ elements/map/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [1.21.3](https://github.com/EOX-A/EOxElements/compare/map-v1.21.2...map-v1.21.3) (2025-06-10)
+
+
+### Bug Fixes
+
+* Min/max zoom was setting the zoom level even when the current min and max level same, which was causing some weird zoom level issue ([#1714](https://github.com/EOX-A/EOxElements/issues/1714)) ([f8cd03d](https://github.com/EOX-A/EOxElements/commit/f8cd03dc61cb9a050e95b0a40d2172dd3ce0e5ef))
+
+## [1.21.2](https://github.com/EOX-A/EOxElements/compare/map-v1.21.1...map-v1.21.2) (2025-05-20)
+
+
+### Bug Fixes
+
+* Disable zoom in/out buttons when reaching limits ([#1633](https://github.com/EOX-A/EOxElements/issues/1633)) ([0ddca87](https://github.com/EOX-A/EOxElements/commit/0ddca87306d076f038a2f7b4978ccde00ea6dbe8))
+
+## [1.21.1](https://github.com/EOX-A/EOxElements/compare/map-v1.21.0...map-v1.21.1) (2025-05-08)
+
+
+### Bug Fixes
+
+* Refactor type `EoxLayer` & Generate Layer JSON Schema ([#1615](https://github.com/EOX-A/EOxElements/issues/1615)) ([34dfb3e](https://github.com/EOX-A/EOxElements/commit/34dfb3e717adcb889612712aebf03a94b17e0d67))
+
 ## [1.21.0](https://github.com/EOX-A/EOxElements/compare/map-v1.20.0...map-v1.21.0) (2025-04-10)